Refactored marl wrapper implementation
