+++ cares_reinforcement_learning/envs/marl/f1tenth/multi_f1tenth_environment.py
@@ -8,34 +8,41 @@ from cares_reinforcement_learning.envs.marl.marl_environment import MARLEnvironm
 from cares_reinforcement_learning.envs.configurations import MultiF1TenthConfig
 from cares_reinforcement_learning.types.experience import MultiAgentExperience
 from cares_reinforcement_learning.types.observation import MARLObservation
+from cares_reinforcement_learning.util import helpers as hlp
 
 
 class F1TenthMARLEnvironment(MARLEnvironment):
     def __init__(self, config: MultiF1TenthConfig, seed: int) -> None:
         super().__init__(config, seed)
 
-        # Import here to avoid a hard ROS2 dependency at module load time
         self.factory = EnvironmentFactory()
-
-        # # Instantiate the task
         self.env = self.factory.create(config.task, {})
 
-        # Stable agent ordering — all list outputs follow this order
-        self.possible_agents: list[str] = self.env.agents
-
+        # Stable agent ordering: all list/dict outputs follow this order.
+        self.possible_agents: list[str] = list(self.env.agents)
+        self.agent_teams = self._split_agents_by_team(self.possible_agents)
+        self.apply_action_normalization = True
         self.observation: MARLObservation
+
+        self.set_seed(seed)
 
     # ------------------------------------------------------------------
     # MARLEnvironment abstract properties
     # ------------------------------------------------------------------
 
     @cached_property
-    def max_action_value(self) -> list[np.ndarray]:
-        return [self.env.max_actions.copy() for _ in self.possible_agents]
+    def max_action_value(self) -> dict[str, np.ndarray]:
+        return {
+            agent: np.asarray(self.env.max_actions, dtype=np.float32).copy()
+            for agent in self.possible_agents
+        }
 
     @cached_property
-    def min_action_value(self) -> list[np.ndarray]:
-        return [self.env.min_actions.copy() for _ in self.possible_agents]
+    def min_action_value(self) -> dict[str, np.ndarray]:
+        return {
+            agent: np.asarray(self.env.min_actions, dtype=np.float32).copy()
+            for agent in self.possible_agents
+        }
 
     @cached_property
     def observation_space(self) -> dict[str, Any]:
@@ -44,7 +51,7 @@ class F1TenthMARLEnvironment(MARLEnvironment):
             "obs": dict(raw["obs"]),  # dict[agent_id -> per-agent obs dim]
             "state": int(raw["state"]),  # int, concatenated global state dim
             "num_agents": int(raw["num_agents"]),
-            "teams": dict(raw["teams"]),
+            "teams": dict(raw.get("teams", self.agent_teams)),
         }
 
     @cached_property
@@ -78,15 +85,27 @@ class F1TenthMARLEnvironment(MARLEnvironment):
     # ------------------------------------------------------------------
 
     def set_seed(self, seed: int) -> None:
-        # F1Tenth runs in ROS2/Gazebo — seeding is a no-op, kept for
-        # interface compliance
         self.seed = seed
+        if hasattr(self.env, "set_seed"):
+            self.env.set_seed(seed)
+
+        for index, agent in enumerate(self.possible_agents):
+            action_space = getattr(self.env, "action_spaces", {}).get(agent)
+            if hasattr(action_space, "seed"):
+                action_space.seed(seed + index)
 
     def sample_action(self) -> dict[str, np.ndarray]:
-        return {
-            agent: self.env.action_spaces[agent].sample()
-            for agent in self.possible_agents
-        }
+        actions = {}
+        for agent in self.possible_agents:
+            action = self.env.action_spaces[agent].sample()
+            if self.apply_action_normalization:
+                action = hlp.normalize(
+                    action,
+                    self.max_action_value[agent],
+                    self.min_action_value[agent],
+                )
+            actions[agent] = np.asarray(action, dtype=np.float32)
+        return actions
 
     def reset(self, training: bool = True) -> MARLObservation:
         obs_dict, _ = self.env.reset()
@@ -100,7 +119,20 @@ class F1TenthMARLEnvironment(MARLEnvironment):
         return self.observation
 
     def step(self, action: dict[str, np.ndarray]) -> MultiAgentExperience:
-        obs_dict, rewards, terminateds, truncateds, infos = self.env.step(action)
+        environment_action = action.copy()
+        if self.apply_action_normalization:
+            environment_action = {
+                agent: hlp.denormalize(
+                    np.asarray(action[agent], dtype=np.float32),
+                    self.max_action_value[agent],
+                    self.min_action_value[agent],
+                )
+                for agent in self.possible_agents
+            }
+
+        obs_dict, rewards, terminateds, truncateds, infos = self.env.step(
+            environment_action
+        )
 
         next_observation = MARLObservation(
             global_state=self._build_global_state(obs_dict),
@@ -110,7 +142,10 @@ class F1TenthMARLEnvironment(MARLEnvironment):
 
         experience = MultiAgentExperience(
             observation=self.observation.clone(),
-            action=action,
+            action={
+                agent: np.asarray(action[agent], dtype=np.float32).copy()
+                for agent in self.possible_agents
+            },
             reward={a: rewards[a] for a in self.possible_agents},
             next_observation=next_observation.clone(),
             done={a: terminateds[a] for a in self.possible_agents},
